modified the setters accordingly

+++ cpp_M03/ex00/main.cpp
@@ -11,8 +11,8 @@ int	main( void )
 	ClapTrap	g( faux_g );
 
 	m.attack( "Gotaga" );
-	g.takeDamage( (unsigned int) m.getAD() );
-	g.beRepaired( (unsigned int) m.getAD() );
+	g.takeDamage( m.getAD() );
+	g.beRepaired( m.getAD() );
 
 	return ( 0 );
 }
